pyproject.toml
[project]
name = "querymancer"
version = "0.1.0"
description = "AI agent that lets you to talk to your database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.21",
    "langchain-groq>=0.3.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
    "watchdog>=6.0.0",
    "sambanova-ai",
    "requests>=2.31.0",  # Required for API calls to SambaNova
]

[dependency-groups]
dev = [
    "faker>=37.1.0",
    "pre-commit>=4.0.1",
    "ruff>=0.8.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.setuptools]
packages = ["querymancer"]
